cleaned

+++ main/src/lights.rs
@@ -1,7 +1,6 @@
 use crossbeam_channel as cbc;
-use log::debug;
-use log::trace;
-use crate::messages;
+use log::{debug, trace};
+use crate::models::messages;
 use crate::fsm;
 use driver_rust::elevio::elev as e;
 use crate::config;
@@ -20,11 +19,12 @@ pub fn run(lights_rx: cbc::Receiver<messages::Controller>, elev_conn: e::Elevato
         }
     }
 }
+
 fn set_all_lights(elev_conn: &e::Elevator, requests: &fsm::ControllerRequests) {
     trace!("set_all_lights");
     for f in 0..config::FLOOR_COUNT {
         for b in 0..config::CALL_COUNT {
-            elev_conn.call_button_light(f as u8, b as u8, requests[f as usize][b as usize]);
+            elev_conn.call_button_light(f as u8, b as u8, requests[f][b]);
         }
     }
 }
